Breaking featuredButton and featuredStatus out into components
